+++ src/init.jl
@@ -1,4 +1,9 @@
+using REPL
+
+const use_local_dependencies = Ref(false)
+
 function __init__()
+
     if isdefined(Main, :PLOTS_DEFAULTS)
         if haskey(Main.PLOTS_DEFAULTS, :theme)
             theme(Main.PLOTS_DEFAULTS[:theme])
@@ -7,20 +12,16 @@ function __init__()
             k == :theme || default(k, v)
         end
     end
-    pushdisplay(PlotsDisplay())
+
+    insert!(Base.Multimedia.displays, findlast(x -> x isa Base.TextDisplay || x isa REPL.REPLDisplay, Base.Multimedia.displays) + 1, PlotsDisplay())
 
     atreplinit(i -> begin
-        if PlotDisplay() in Base.Multimedia.displays
+        while PlotsDisplay() in Base.Multimedia.displays
             popdisplay(PlotsDisplay())
         end
-        pushdisplay(PlotsDisplay())
+        insert!(Base.Multimedia.displays, findlast(x -> x isa REPL.REPLDisplay, Base.Multimedia.displays) + 1, PlotsDisplay())
     end)
 
-    include(joinpath(@__DIR__, "backends", "plotly.jl"))
-    include(joinpath(@__DIR__, "backends", "gr.jl"))
-    include(joinpath(@__DIR__, "backends", "web.jl"))
-
-    @require GLVisualize = "4086de5b-f4b6-55f3-abb0-b8c73827585f" include(joinpath(@__DIR__, "backends", "glvisualize.jl"))
     @require HDF5 = "f67ccb44-e63f-5c2f-98bd-6dc0ccc4ba2f" include(joinpath(@__DIR__, "backends", "hdf5.jl"))
     @require InspectDR = "d0351b0e-4b05-5898-87b3-e2a8edfddd1d" include(joinpath(@__DIR__, "backends", "inspectdr.jl"))
     @require PGFPlots = "3b7a836e-365b-5785-a47d-02c71176b4aa" include(joinpath(@__DIR__, "backends", "pgfplots.jl"))
@@ -31,10 +32,11 @@ function __init__()
     # ---------------------------------------------------------
     # IJulia
     # ---------------------------------------------------------
-
+    use_local = false
     @require IJulia = "7073ff75-c697-5162-941a-fcdaad2a7d2a" begin
         if IJulia.inited
-
+            # IJulia is more stable with local file
+            use_local = isfile(plotly_local_file_path)
             """
             Add extra jupyter mimetypes to display_dict based on the plot backed.
 
@@ -82,6 +84,40 @@ function __init__()
             end
 
             ENV["MPLBACKEND"] = "Agg"
+        end
+    end
+
+    if haskey(ENV, "PLOTS_HOST_DEPENDENCY_LOCAL")
+        use_local = ENV["PLOTS_HOST_DEPENDENCY_LOCAL"] == "true"
+        use_local_dependencies[] = isfile(plotly_local_file_path) && use_local
+        if use_local && !isfile(plotly_local_file_path)
+            @warn("PLOTS_HOST_DEPENDENCY_LOCAL is set to true, but no local plotly file found. run Pkg.build(\"Plots\") and make sure PLOTS_HOST_DEPENDENCY_LOCAL is set to true")
+        end
+    else
+        use_local_dependencies[] = use_local
+    end
+
+
+
+    # ---------------------------------------------------------
+    # A backup, if no PNG generation is defined, is to try to make a PDF and use FileIO to convert
+    @require FileIO = "5789e2e9-d7fb-5bc7-8068-2c6fae9b9549" begin
+        PDFBackends = Union{PGFPlotsBackend,PlotlyJSBackend,PyPlotBackend,InspectDRBackend,GRBackend}
+        function _show(io::IO, ::MIME"image/png", plt::Plot{<:PDFBackends})
+            fn = tempname()
+
+            # first save a pdf file
+            pdf(plt, fn)
+
+            # load that pdf into a FileIO Stream
+            s = FileIO.load(fn * ".pdf")
+
+            # save a png
+            pngfn = fn * ".png"
+            FileIO.save(pngfn, s)
+
+            # now write from the file
+            write(io, read(open(pngfn), String))
         end
     end
 end